requirements.txt
requests
beautifulsoup4
openai
streamlit
pytest
google-search-results